Extracted organization twitter count shared example
